CRS-2131 Fixing issues with unused adas. (#851)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/service/BookingExtractionService.kt
@@ -650,7 +650,7 @@ class BookingExtractionService(
       rules = setOf(calculationRule),
       releaseDate = adjustedReleaseDate,
       unadjustedDate = hdcedSentenceDate,
-      adjustedDays = ChronoUnit.DAYS.between(adjustedReleaseDate, hdcedSentenceDate).toInt(),
+      adjustedDays = ChronoUnit.DAYS.between(adjustedReleaseDate, hdcedSentenceDate),
     )
   }
